pyproject.toml
[project]
name = "housing-forecast-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.21",
    "mcp>=1.13.1",
    "strands-agents>=1.6.0",
    "strands-agents-tools>=0.2.5",
    "streamlit>=1.49.1",
]
